fix movies for 64-bit machines

+++ tools/PwMoviePlayer/GLWidget.h
@@ -1,6 +1,8 @@
 #ifndef GLWidget_h
 #define GLWidget_h
 
+#include <stdint.h>
+
 #include <QGLWidget>
 
 #include "PwMovieUtils.h"
@@ -10,8 +12,8 @@ class GLWidget : public QGLWidget
 	Q_OBJECT
 
 public:
-	GLWidget( QWidget *parent = 0, unsigned long widthParam = 400, unsigned long heightParam = 400, unsigned long movieVersionParam = kCurrentMovieVersion,
-				FILE* movieFileParam = NULL, char** legendParam = NULL, unsigned long endFrameParam = 0, double frameRateParam = 0.0 );
+	GLWidget( QWidget *parent = 0, uint32_t widthParam = 400, uint32_t heightParam = 400, uint32_t movieVersionParam = kCurrentMovieVersion,
+				FILE* movieFileParam = NULL, char** legendParam = NULL, uint32_t endFrameParam = 0, double frameRateParam = 0.0 );
 	~GLWidget();
 	
 	void Draw();
@@ -25,19 +27,19 @@ protected:
 	void resizeGL( int width, int height );
 
 private:
-	unsigned long	width;
-	unsigned long	height;
-	unsigned long	movieVersion;
+	uint32_t	width;
+	uint32_t	height;
+	uint32_t	movieVersion;
 	FILE*			movieFile;
 	char**			legend;
-	unsigned long	endFrame;
+	uint32_t	endFrame;
 	double			frameRate;
 
-	ulong		rgbBufSize;
-	ulong		rleBufSize;
-	ulong*		rgbBuf1;	// actually allocated
-//	ulong*		rgbBuf2;	// actually allocated
-	ulong*		rleBuf;
+	uint32_t		rgbBufSize;
+	uint32_t		rleBufSize;
+	uint32_t*		rgbBuf1;	// actually allocated
+//	uint32_t*		rgbBuf2;	// actually allocated
+	uint32_t*		rleBuf;
 };
 
 #endif
